+++ src/main/java/com/baloise/cc2023diabetes/service/food/model/CreateRecipeIngredientModel.java
@@ -2,8 +2,6 @@ package com.baloise.cc2023diabetes.service.food.model;
 
 import io.soabase.recordbuilder.core.RecordBuilder;
 
-import java.math.BigDecimal;
-
 @RecordBuilder
-public record CreateRecipeIngredientModel(Integer foodId, BigDecimal amount) {
+public record CreateRecipeIngredientModel(Long foodId, Double amount) {
 }
