+++ src/main/java/com/baloise/cc2023diabetes/service/food/model/FoodModel.java
@@ -1,11 +1,10 @@
 package com.baloise.cc2023diabetes.service.food.model;
 
-import java.math.BigDecimal;
-
-public record FoodModel(Integer id, String title,
+public record FoodModel(Long id,
+                        String title,
                         String unit,
-                        BigDecimal carbohydrateAvailable,
-                        BigDecimal carbohydrateUnits,
-                        BigDecimal sugar) {
+                        Double carbohydrateAvailable,
+                        Double carbohydrateUnits,
+                        Double sugar) {
 
 }
