+++ src/main/java/com/baloise/cc2023diabetes/service/food/model/RecipeIngredientModel.java
@@ -1,7 +1,5 @@
 package com.baloise.cc2023diabetes.service.food.model;
 
-import java.math.BigDecimal;
-
 public record RecipeIngredientModel(FoodModel food,
-                                    BigDecimal amount) {
+                                    Double amount) {
 }
